Make GitHub auth credentials optional (for public repos)
When reading info and fetching releases from a public GH repo, there's no need for an access token. This commit makes it so that we leave off the Authorization header and HTTP auth fields when no token nor username & password are provided.

Tested against a public repo and Electron's autoupdater.